Repository取得をPresenterに任せる
